refactor: Extract requireAdmin helper and rename shadowing fetch variable
