+++ apps/web/src/app/api/u/[slug]/kb-github-remote/repos/route.ts
@@ -4,24 +4,9 @@ import { auditEvent } from '@/lib/auth'
 import { getInstallationRepos } from '@/lib/git/github-app-auth'
 import type { KbGithubRemoteRepo } from '@/lib/kb-github-remote/types'
 import { kbGithubRemoteService } from '@/lib/services'
-import { requireCapability } from '@/lib/runtime/require-capability'
 import { withAuth } from '@/lib/runtime/with-auth'
 
-function requireAdmin(user: { id: string; role: string }) {
-  const denied = requireCapability('kbGithubSync')
-  if (denied) {
-    return { ok: false as const, response: denied }
-  }
-
-  if (user.role !== 'ADMIN') {
-    return {
-      ok: false as const,
-      response: NextResponse.json({ error: 'forbidden' }, { status: 403 }),
-    }
-  }
-
-  return { ok: true as const }
-}
+import { requireAdmin } from '../require-admin'
 
 export const GET = withAuth<{ repos: KbGithubRemoteRepo[] } | { error: string }>(
   { csrf: false },
