+++ apps/web/src/app/api/u/[slug]/kb-github-remote/route.ts
@@ -3,24 +3,9 @@ import { NextResponse } from 'next/server'
 import { auditEvent } from '@/lib/auth'
 import type { KbGithubRemoteIntegrationGetResponse } from '@/lib/kb-github-remote/types'
 import { kbGithubRemoteService } from '@/lib/services'
-import { requireCapability } from '@/lib/runtime/require-capability'
 import { withAuth } from '@/lib/runtime/with-auth'
 
-function requireAdmin(user: { id: string; role: string }) {
-  const denied = requireCapability('kbGithubSync')
-  if (denied) {
-    return { ok: false as const, response: denied }
-  }
-
-  if (user.role !== 'ADMIN') {
-    return {
-      ok: false as const,
-      response: NextResponse.json({ error: 'forbidden' }, { status: 403 }),
-    }
-  }
-
-  return { ok: true as const }
-}
+import { requireAdmin } from './require-admin'
 
 function serializeResponse(
   record: Awaited<ReturnType<typeof kbGithubRemoteService.findIntegration>>,
