+++ apps/web/src/app/api/u/[slug]/kb-github-remote/setup/route.ts
@@ -3,24 +3,9 @@ import { NextRequest, NextResponse } from 'next/server'
 import { auditEvent } from '@/lib/auth'
 import { exchangeManifestCode } from '@/lib/git/github-app-auth'
 import { kbGithubRemoteService } from '@/lib/services'
-import { requireCapability } from '@/lib/runtime/require-capability'
 import { withAuth } from '@/lib/runtime/with-auth'
 
-function requireAdmin(user: { id: string; role: string }) {
-  const denied = requireCapability('kbGithubSync')
-  if (denied) {
-    return { ok: false as const, response: denied }
-  }
-
-  if (user.role !== 'ADMIN') {
-    return {
-      ok: false as const,
-      response: NextResponse.json({ error: 'forbidden' }, { status: 403 }),
-    }
-  }
-
-  return { ok: true as const }
-}
+import { requireAdmin } from '../require-admin'
 
 export const GET = withAuth<{ error: string }>(
   { csrf: false },
