+++ apps/web/src/app/api/u/[slug]/kb-github-remote/sync/route.ts
@@ -4,24 +4,9 @@ import { auditEvent } from '@/lib/auth'
 import { pushToGithub, pullFromGithub } from '@/lib/git/kb-github-sync'
 import type { KbGithubRemoteSyncState } from '@/lib/services/kb-github-remote'
 import { kbGithubRemoteService } from '@/lib/services'
-import { requireCapability } from '@/lib/runtime/require-capability'
 import { withAuth } from '@/lib/runtime/with-auth'
 
-function requireAdmin(user: { id: string; role: string }) {
-  const denied = requireCapability('kbGithubSync')
-  if (denied) {
-    return { ok: false as const, response: denied }
-  }
-
-  if (user.role !== 'ADMIN') {
-    return {
-      ok: false as const,
-      response: NextResponse.json({ error: 'forbidden' }, { status: 403 }),
-    }
-  }
-
-  return { ok: true as const }
-}
+import { requireAdmin } from '../require-admin'
 
 export const POST = withAuth(
   { csrf: true },
