+++ apps/web/src/app/api/u/[slug]/kb-github-remote/test/route.ts
@@ -2,24 +2,9 @@ import { NextResponse } from 'next/server'
 
 import { getInstallationToken } from '@/lib/git/github-app-auth'
 import { kbGithubRemoteService } from '@/lib/services'
-import { requireCapability } from '@/lib/runtime/require-capability'
 import { withAuth } from '@/lib/runtime/with-auth'
 
-function requireAdmin(user: { id: string; role: string }) {
-  const denied = requireCapability('kbGithubSync')
-  if (denied) {
-    return { ok: false as const, response: denied }
-  }
-
-  if (user.role !== 'ADMIN') {
-    return {
-      ok: false as const,
-      response: NextResponse.json({ error: 'forbidden' }, { status: 403 }),
-    }
-  }
-
-  return { ok: true as const }
-}
+import { requireAdmin } from '../require-admin'
 
 export const POST = withAuth<{ ok: boolean; message?: string } | { error: string }>(
   { csrf: true },
